IOSS: Sync from sierra
Primarily formatting changes using clang-format to get consistent
source structure over code-base.

Switch all std::endl to '\n'

Move Ioss_GetLongOpt into Ioss library

clean up some compilation warnings.

(From STK/SEACAS)

+++ packages/seacas/libraries/ioss/src/cgns/Iocgns_DecompositionData.C
@@ -699,7 +699,7 @@ namespace Iocgns {
       // * Other method uses 6*ioss_node_count extra memory; 3 reads;
       // and 1 communicate_node_data call.
       //
-      for (size_t d = 0; d < m_decomposition.m_spatialDimension; d++) {
+      for (int d = 0; d < m_decomposition.m_spatialDimension; d++) {
         get_file_node_coordinates(filePtr, d, TOPTR(tmp));
         communicate_node_data(TOPTR(tmp), TOPTR(ioss_tmp), 1);
